@@ -18,48 +18,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
+    "@angular/animations": "4.3.3",
     "@angular/cli": "1.2.7",
-    "@angular/common": "4.0.0",
-    "@angular/compiler": "4.0.0",
-    "@angular/compiler-cli": "4.0.0",
-    "@angular/core": "4.0.0",
-    "@angular/forms": "4.0.0",
-    "@angular/http": "4.0.0",
-    "@angular/platform-browser": "4.0.0",
-    "@angular/platform-browser-dynamic": "4.0.0",
-    "@angular/router": "4.0.0",
+    "@angular/common": "4.3.3",
+    "@angular/compiler": "4.3.3",
+    "@angular/compiler-cli": "4.3.3",
+    "@angular/core": "4.3.3",
+    "@angular/forms": "4.3.3",
+    "@angular/platform-browser": "4.3.3",
+    "@angular/platform-browser-dynamic": "4.3.3",
+    "@angular/router": "4.3.3",
     "admin-lte": "2.3.11",
     "core-js": "2.4.1",
-    "express": "^4.16.4",
     "font-awesome": "4.7.0",
     "intl": "1.2.5",
     "jquery": "3.1.1",
-    "json-server": "^0.14.0",
-    "path": "^0.12.7",
-    "reflect-metadata": "0.1.9",
-    "rxjs": "5.1.0",
+    "reflect-metadata": "0.1.10",
+    "rxjs": "5.4.2",
     "ts-helpers": "1.1.2",
-    "typescript": "2.2.0",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "0.8.4"
+    "typescript": "2.4.2",
+    "web-animations-js": "2.2.5",
+    "zone.js": "0.8.16"
   },
   "devDependencies": {
-    "@types/jasmine": "2.5.38",
+    "@types/jasmine": "2.5.53",
+    "@types/express": "4.0.37",
+    "@types/jsonwebtoken": "7.2.3",
     "@types/node": "7.0.5",
-    "codelyzer": "2.0.0",
-    "enhanced-resolve": "^3.3.0",
-    "har-validator": "^5.1.3",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "3.2.0",
-    "karma": "1.4.1",
-    "karma-chrome-launcher": "2.0.0",
+    "codelyzer": "3.1.2",
+    "jasmine-core": "2.7.0",
+    "jasmine-spec-reporter": "4.1.1",
+    "json-server": "0.12.0",
+    "jsonwebtoken": "7.4.1",
+    "karma": "1.7.0",
+    "karma-chrome-launcher": "2.2.0",
     "karma-cli": "1.0.1",
     "karma-jasmine": "1.1.0",
-    "karma-remap-istanbul": "0.2.2",
-    "protractor": "5.1.0",
-    "ts-node": "2.1.0",
-    "tslint": "4.5.0",
-    "webdriver-manager": "10.2.5"
+    "karma-remap-istanbul": "0.6.0",
+    "protractor": "5.1.2",
+    "ts-node": "3.3.0",
+    "tslint": "5.5.0",
+    "webdriver-manager": "12.0.6"
   }
 }
