@@ -3,13 +3,16 @@ import { ShoppingCartService } from "app/restaurant-detail/shopping-cart/shoppin
 import { CartItem } from "app/restaurant-detail/shopping-cart/cart-item.model";
 import { Order } from "./order.model";
 import { Observable } from "rxjs/Observable";
-import { Http, Headers, RequestOptions } from "@angular/http";
+import { HttpClient, HttpHeaders} from "@angular/common/http";
 import { MEAT_API } from '../app.api';
+import { LoginService } from '../security/login/login.service';
 
 @Injectable()
 export class OrderService {
 
-    constructor(private cartService: ShoppingCartService, private http: Http){}
+    constructor(private cartService: ShoppingCartService, 
+                private http: HttpClient, 
+                private loginService: LoginService){}
 
     cartItems(): CartItem[]{
         return this.cartService.items;
@@ -32,12 +35,13 @@ export class OrderService {
     }
 
     checkOrder(order: Order): Observable<string> {
-        const headers = new Headers();
-        headers.append('Content-Type', 'application/json')
-        return this.http.post(`${MEAT_API}/order`, 
-                                JSON.stringify(order),
-                                 new RequestOptions({headers: headers}))
-                        .map(response => response.json());
+        let headers = new HttpHeaders()
+        if (this.loginService.isUserLoggedIn) {
+            headers = headers.set('Authorization', `Bearer ${this.loginService.user.accessToken}`)
+            console.log(this.loginService.user.accessToken)
+        }
+        return this.http.post<Order>(`${MEAT_API}/api/order`, order, {headers: headers})
+                                    .map(order => order._id);
     }
 
     clear() {
